+++ src/index.ts
@@ -1,13 +1,11 @@
-export { default as display } from "./display";
+export { display } from "./display.js";
 
 // Switch back to one line like the others after TypeScript 3.8, see below
-import generate from "./generate";
+import { generate } from "./generate.js";
 export { generate };
 
 // Usually not needed, but just in case...
-export { default as svgs } from "./svgs";
-export { default as svgsIndex } from "./svgs-index";
+export { default as svgs } from "./svgs.js";
+export { svgsIndex } from "./svgs-index.js";
 
-// Uncomment this when TypeScript 3.8 is out. Otherwise importing and re-exporting breaks
-// export type { Face } from "./generate";
-export type Face = ReturnType<typeof generate>;
+export type { Face } from "./generate.js";